robodriver/robots/robodriver-robot-realman-aio-ros1/robodriver_robot_realman_aio_ros1/config.py
from typing import Dict
from dataclasses import dataclass, field

from lerobot.robots.config import RobotConfig
from lerobot.cameras import CameraConfig
from lerobot.cameras.opencv import OpenCVCameraConfig
from lerobot.motors import Motor, MotorNormMode


@RobotConfig.register_subclass("realman_aio_ros1")
@dataclass
class RealmanAioRos1RobotConfig(RobotConfig):
    use_degrees = True
    norm_mode_body = (
        MotorNormMode.DEGREES if use_degrees else MotorNormMode.RANGE_M100_100
    )

    # 按组件分组：{ comp_id: { joint_name: Motor, ... }, ... }
    leader_motors: Dict[str, Dict[str, Motor]] = field(
        default_factory=lambda norm_mode_body=norm_mode_body: {
            "leader_left_joints": {
                "joint_0": Motor(1, "robot_motor", norm_mode_body),
                "joint_1": Motor(2, "robot_motor", norm_mode_body),
                "joint_2": Motor(3, "robot_motor", norm_mode_body),
                "joint_3": Motor(4, "robot_motor", norm_mode_body),
                "joint_4": Motor(5, "robot_motor", norm_mode_body),
                "joint_5": Motor(6, "robot_motor", norm_mode_body),
                "joint_6": Motor(7, "robot_motor", norm_mode_body),
            },
            "leader_right_joints": {
                "joint_0": Motor(1, "robot_motor", norm_mode_body),
                "joint_1": Motor(2, "robot_motor", norm_mode_body),
                "joint_2": Motor(3, "robot_motor", norm_mode_body),
                "joint_3": Motor(4, "robot_motor", norm_mode_body),
                "joint_4": Motor(5, "robot_motor", norm_mode_body),
                "joint_5": Motor(6, "robot_motor", norm_mode_body),
                "joint_6": Motor(7, "robot_motor", norm_mode_body),
            },
        }
    )

    follower_motors: Dict[str, Dict[str, Motor]] = field(
        default_factory=lambda norm_mode_body=norm_mode_body: {
            "left_arm": {
                "position_x": Motor(1, "robot_motor", norm_mode_body),
                "position_y": Motor(2, "robot_motor", norm_mode_body),
                "position_z": Motor(3, "robot_motor", norm_mode_body),
                "orientation_x": Motor(4, "robot_motor", norm_mode_body),
                "orientation_y": Motor(5, "robot_motor", norm_mode_body),
                "orientation_z": Motor(6, "robot_motor", norm_mode_body),
                "orientation_w": Motor(7, "robot_motor", norm_mode_body),
            },
            "right_arm": {
                "position_x": Motor(1, "robot_motor", norm_mode_body),
                "position_y": Motor(2, "robot_motor", norm_mode_body),
                "position_z": Motor(3, "robot_motor", norm_mode_body),
                "orientation_x": Motor(4, "robot_motor", norm_mode_body),
                "orientation_y": Motor(5, "robot_motor", norm_mode_body),
                "orientation_z": Motor(6, "robot_motor", norm_mode_body),
                "orientation_w": Motor(7, "robot_motor", norm_mode_body),
            },
            "follower_left_joints": {
                "joint_0": Motor(1, "robot_motor", norm_mode_body),
                "joint_1": Motor(2, "robot_motor", norm_mode_body),
                "joint_2": Motor(3, "robot_motor", norm_mode_body),
                "joint_3": Motor(4, "robot_motor", norm_mode_body),
                "joint_4": Motor(5, "robot_motor", norm_mode_body),
                "joint_5": Motor(6, "robot_motor", norm_mode_body),
                "joint_6": Motor(7, "robot_motor", norm_mode_body),
            },
            "follower_right_joints": {
                "joint_0": Motor(1, "robot_motor", norm_mode_body),
                "joint_1": Motor(2, "robot_motor", norm_mode_body),
                "joint_2": Motor(3, "robot_motor", norm_mode_body),
                "joint_3": Motor(4, "robot_motor", norm_mode_body),
                "joint_4": Motor(5, "robot_motor", norm_mode_body),
                "joint_5": Motor(6, "robot_motor", norm_mode_body),
                "joint_6": Motor(7, "robot_motor", norm_mode_body),
            },
        }
    )

    cameras: Dict[str, CameraConfig] = field(
        default_factory=lambda: {
            "image_top": OpenCVCameraConfig(index_or_path=0, fps=30, width=640, height=480),
            "image_left": OpenCVCameraConfig(index_or_path=1, fps=30, width=640, height=480),
            "image_right": OpenCVCameraConfig(index_or_path=2, fps=30, width=640, height=480),
        }
    )

    use_videos: bool = False

    microphones: Dict[str, int] = field(default_factory=lambda: {}
    )
