robodriver/robots/robodriver-robot-realman-aio-ros1/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robodriver_robot_realman_aio_ros1"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
keywords = ["robotics", "lerobot", "realman_aio_ros1"]

dependencies = [
    "logging_mp",
    "numpy",
    "pyzmq",
    "opencv-python",
]

[tool.setuptools.packages.find]
include = ["robodriver_robot_realman_aio_ros1"]

